requirements.txt
# Core dependencies
streamlit==1.28.0
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1

# API and Web
fastapi==0.104.1
uvicorn==0.23.2
websockets==11.0.3
httpx==0.25.0

# Data sources
yfinance==0.2.31
pandas-datareader==0.10.0

# Machine learning
scikit-learn==1.3.2
xgboost==2.0.1
lightgbm==4.1.0
torch==2.1.0
transformers==4.34.1

# Visualization
plotly==5.17.0
altair==5.1.2

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3.post1
tqdm==4.66.1

# Performance optimization
joblib==1.3.2
numba==0.58.1
cachetools==5.3.2